duk_substring() unicode support implemented

+++ src/duk_api.h
@@ -434,7 +434,7 @@ void duk_concat(duk_context *ctx, unsigned int count);              /* [val1 ...
 void duk_join(duk_context *ctx, unsigned int count);                /* [sep val1 ... valN] -> [res], coerced and joined */
 void duk_decode_string(duk_context *ctx, int index, duk_decode_char_function callback, void *udata);
 void duk_map_string(duk_context *ctx, int index, duk_map_char_function callback, void *udata);
-void duk_substring(duk_context *ctx, unsigned int start_offset, unsigned int end_offset);
+void duk_substring(duk_context *ctx, size_t start_offset, size_t end_offset);
 
 /*
  *  Ecmascript operators
